requirements.txt
absl-py==2.3.0
accelerate==1.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
bitsandbytes==0.35.4
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
decord==0.6.0
diffusers==0.14.0
einops==0.8.1
executing==2.2.0
filelock==3.13.1
fsspec==2024.6.1
ftfy==6.3.1
grpcio==1.73.0
hf-xet==1.1.3
huggingface-hub==0.25.2
icecream==2.1.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.4
Markdown==3.8
MarkupSafe==2.1.5
modelcards==0.1.6
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pillow==11.0.0
protobuf==6.31.1
psutil==7.0.0
Pygments==2.19.1
pytorch-triton==3.3.1+gitc8757738
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
six==1.17.0
sympy==1.13.3
tokenizers==0.13.3
tqdm==4.67.1
transformers==4.31.0
triton==3.3.0
typing_extensions==4.12.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.23.0
